+++ common/src/main/java/io/github/lolens/showcaser/registry/ShareHandlerRegistrar.java
@@ -54,7 +54,6 @@ public final class ShareHandlerRegistrar {
         Showcaser.LOGGER.debug("Registered server share handler with id {}", serverShareHandler.getIdentifier());
     }
 
-    // todo forgot priority......
     public static <T extends Screen> void registerClientSummary(ClientSummary<T> summary) {
 
         if (summary.contextCreator() != null) {